pyproject.toml
 [tool.poetry]
name = "geodatacrawler"
version = "0.1.0"
description = ""
authors = ["Paul van Genuchten <paul.vangenuchten@wur.nl>"]
packages = [
    { include = "geodatacrawler", from = "src" },
    { include = "geodatacrawler/templates/*", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
Fiona = "^1.8.21"
rasterio = "^1.2.10"
pysqlite3 = "^0.4.7"
mappyfile = "^0.9.7"
GDAL = "3.3.2"
pygeometa = "^0.11.1"
OWSLib = "^0.26.0"
pyproj = "3.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
crawl-metadata = "geodatacrawler.metadata:indexDir"
crawl-maps = "geodatacrawler.mapfile:mapForDir"
crawl-etl = "geodatacrawler.etl:etl"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
